Обновил скрипт добавления вещей

+++ SurviveTheCovid/Assets/Scripts/Stats/Stat.cs
@@ -5,24 +5,24 @@ using UnityEngine;
 [System.Serializable]   
 public class Stat
 {
-	[SerializeField] private int baseValue;
+	[SerializeField] private float baseValue;
 
-	private List<int> modifiers = new List<int>();
+	private List<float> modifiers = new List<float>();
 
-	public int GetValue()
+	public float GetValue()
 	{
-		int finalValue = baseValue;
+		float finalValue = baseValue;
 		modifiers.ForEach(m => finalValue += m);
 		return finalValue;
 	}
 
-	public void AddModifiers(int modifier)
+	public void AddModifiers(float modifier)
 	{
 		if (modifier != 0)
 			modifiers.Add(modifier);
 	}
 
-	public void DeleteModifier(int modifier)
+	public void DeleteModifier(float modifier)
 	{
 		if (modifier != 0)
 			modifiers.Remove(modifier);
